refactor: create callback context and code split
