+++ src/hooks/useAsync/constants.ts
@@ -1,0 +1,9 @@
+import { UseAsyncState } from './types';
+
+export const CANCEL_SYMBOL = Symbol('CANCELLED');
+
+export const INITIAL_DATA: UseAsyncState = {
+  loading: true,
+  data: undefined,
+  error: undefined,
+};
